Add files via upload
Myocardial infarction project has been my internship project to successfully complete and obtain a certificate at Dr. Hosseinpour's Academy.
The next point is that this data set was a completely Iranian data that we were able to examine in the middle of 5 hospitals where myocardial infarction is high, like a doctor.